Fix global context for non browser environments
Bug:

If we ever try to run the bundle generated by this script in an
environment like NodeJS, we will get a "window is not defined" error.

The `window` variable/context may not be present in environments outside
of the browser, in this case, the Global Object is referenced by the
`global` variable, and not `window`.

Solution:

Fall back to the `global` variable, if `window` is not present.